refactor: tiles.rs splitted in behavior, gutters, render & tile_actions
