+++ libs/elodin-editor/src/ui/tiles/gutters.rs
@@ -1,0 +1,392 @@
+use super::*;
+use egui::{Color32, Stroke};
+
+// Compute sidebar gutter interactions and return share updates to apply.
+pub(super) fn collect_sidebar_gutter_updates(
+    tree: &mut egui_tiles::Tree<Pane>,
+    ui: &mut egui::Ui,
+    painter: egui::Painter,
+    gutter_width: f32,
+    mask_state: &mut SidebarMaskState,
+) -> Vec<ShareUpdate> {
+    struct GutterCtx<'a> {
+        tree: &'a mut egui_tiles::Tree<Pane>,
+        ui: &'a mut egui::Ui,
+        painter: egui::Painter,
+        gutter_width: f32,
+        mask_state: &'a mut SidebarMaskState,
+        share_updates: Vec<ShareUpdate>,
+    }
+
+    impl<'a> GutterCtx<'a> {
+        fn sidebar_kind(&self, id: TileId) -> Option<SidebarKind> {
+            match self.tree.tiles.get(id) {
+                Some(Tile::Pane(Pane::Hierarchy)) => Some(SidebarKind::Hierarchy),
+                Some(Tile::Pane(Pane::Inspector)) => Some(SidebarKind::Inspector),
+                Some(Tile::Container(Container::Tabs(tabs))) => {
+                    for child in tabs.children.iter() {
+                        if let Some(kind) = self.sidebar_kind(*child) {
+                            return Some(kind);
+                        }
+                    }
+                    None
+                }
+                Some(Tile::Container(Container::Linear(linear))) => {
+                    for child in linear.children.iter() {
+                        if let Some(kind) = self.sidebar_kind(*child) {
+                            return Some(kind);
+                        }
+                    }
+                    None
+                }
+                Some(Tile::Container(Container::Grid(grid))) => {
+                    for child in grid.children() {
+                        if let Some(kind) = self.sidebar_kind(*child) {
+                            return Some(kind);
+                        }
+                    }
+                    None
+                }
+                _ => None,
+            }
+        }
+
+        fn apply_shares(
+            &mut self,
+            container_id: TileId,
+            left_id: TileId,
+            right_id: TileId,
+            left: f32,
+            right: f32,
+        ) {
+            if let Some(Tile::Container(Container::Linear(linear))) =
+                self.tree.tiles.get_mut(container_id)
+            {
+                linear.shares.set_share(left_id, left);
+                linear.shares.set_share(right_id, right);
+            }
+            self.share_updates
+                .push((container_id, left_id, right_id, left, right));
+            self.ui.ctx().request_repaint();
+        }
+
+        fn compute_sidebar_shares(
+            &self,
+            pair_sum: f32,
+            min_other_share: f32,
+            min_sidebar_share: f32,
+            target_sidebar_share: f32,
+            sidebar_on_left: bool,
+        ) -> (f32, f32) {
+            let max_sidebar_share = (pair_sum - min_other_share).max(0.01);
+            let clamped = target_sidebar_share
+                .max(min_sidebar_share.max(0.01))
+                .min(max_sidebar_share);
+            let left_share = if sidebar_on_left {
+                clamped
+            } else {
+                pair_sum - clamped
+            };
+            let right_share = pair_sum - left_share;
+            (left_share.max(0.01), right_share.max(0.01))
+        }
+
+        fn process_pair(
+            &mut self,
+            container_id: TileId,
+            parent_rect: egui::Rect,
+            left_id: TileId,
+            left_rect: egui::Rect,
+            left_kind: Option<SidebarKind>,
+            right_id: TileId,
+            right_rect: egui::Rect,
+            right_kind: Option<SidebarKind>,
+        ) {
+            let left_sidebar = left_kind.is_some();
+            let right_sidebar = right_kind.is_some();
+            if left_sidebar == right_sidebar {
+                return;
+            }
+            let sidebar_kind = match (left_kind, right_kind) {
+                (Some(k), _) => k,
+                (_, Some(k)) => k,
+                _ => return,
+            };
+            let sidebar_on_left = left_sidebar;
+            let pair_width = left_rect.width() + right_rect.width();
+            let (share_left, share_right) = match self.tree.tiles.get(container_id) {
+                Some(Tile::Container(Container::Linear(linear))) => {
+                    (linear.shares[left_id], linear.shares[right_id])
+                }
+                _ => return,
+            };
+            let pair_sum = share_left + share_right;
+            if pair_sum <= 0.0 {
+                return;
+            }
+            let share_per_px = if pair_width > 0.0 {
+                pair_sum / pair_width
+            } else {
+                0.0
+            };
+            let min_sidebar_px = self.gutter_width + 4.0;
+            let min_other_px = 32.0;
+            let min_sidebar_share = if share_per_px > 0.0 {
+                min_sidebar_px * share_per_px
+            } else {
+                pair_sum * 0.05
+            };
+            let min_other_share = if share_per_px > 0.0 {
+                min_other_px * share_per_px
+            } else {
+                pair_sum * 0.05
+            };
+
+            let gap = right_rect.min.x - left_rect.max.x;
+            let mut center_x = (left_rect.max.x + right_rect.min.x) * 0.5;
+            if gap < self.gutter_width {
+                let offset = (self.gutter_width - gap).max(0.0) * 0.5;
+                if left_sidebar {
+                    center_x -= offset;
+                } else {
+                    center_x += offset;
+                }
+            }
+            let half = self.gutter_width * 0.5;
+            center_x = center_x
+                .max(parent_rect.left() + half)
+                .min(parent_rect.right() - half);
+            let gutter_rect = egui::Rect::from_min_max(
+                egui::pos2(center_x - half, parent_rect.top()),
+                egui::pos2(center_x + half, parent_rect.bottom()),
+            );
+
+            let gutter_color = Color32::from_gray(80);
+            let fill = gutter_color;
+            let stroke = Stroke::new(1.0, gutter_color);
+            self.painter.rect_filled(gutter_rect, 0.0, fill);
+            self.painter
+                .rect_stroke(gutter_rect, 0.0, stroke, egui::StrokeKind::Inside);
+
+            let id = self
+                .ui
+                .id()
+                .with(("sidebar_gutter", container_id, left_id, right_id));
+            #[derive(Clone, Copy, Default)]
+            struct DragState {
+                left_width: f32,
+                right_width: f32,
+                start_x: f32,
+                active: bool,
+            }
+            let mut drag_state = self
+                .ui
+                .ctx()
+                .data(|d| d.get_temp::<DragState>(id))
+                .unwrap_or_default();
+
+            let response = self
+                .ui
+                .interact(gutter_rect, id, egui::Sense::click_and_drag())
+                .on_hover_cursor(egui::CursorIcon::PointingHand);
+
+            if response.hovered() {
+                self.ui
+                    .output_mut(|o| o.cursor_icon = egui::CursorIcon::PointingHand);
+            }
+
+            let mut sidebar_masked = self.mask_state.masked(sidebar_kind);
+
+            let click_inside_gutter = response.clicked_by(egui::PointerButton::Primary)
+                && self
+                    .ui
+                    .input(|i| i.pointer.interact_pos())
+                    .map(|p| gutter_rect.shrink(1.0).contains(p))
+                    .unwrap_or(false);
+
+            if click_inside_gutter {
+                let (left_share, right_share) = if sidebar_masked {
+                    let default_px = (parent_rect.width() * 0.15).max(min_sidebar_px);
+                    let restore_share = if share_per_px > 0.0 {
+                        default_px * share_per_px
+                    } else {
+                        pair_sum * 0.15
+                    };
+                    self.compute_sidebar_shares(
+                        pair_sum,
+                        min_other_share,
+                        min_sidebar_share,
+                        restore_share,
+                        sidebar_on_left,
+                    )
+                } else {
+                    let current_sidebar_share = if sidebar_on_left {
+                        share_left
+                    } else {
+                        share_right
+                    };
+                    self.mask_state
+                        .set_last_share(sidebar_kind, Some(current_sidebar_share));
+                    self.compute_sidebar_shares(
+                        pair_sum,
+                        min_other_share,
+                        min_sidebar_share,
+                        min_sidebar_share,
+                        sidebar_on_left,
+                    )
+                };
+                self.apply_shares(container_id, left_id, right_id, left_share, right_share);
+                self.mask_state.set_masked(sidebar_kind, !sidebar_masked);
+                sidebar_masked = !sidebar_masked;
+            }
+
+            if sidebar_masked {
+                if drag_state.active {
+                    self.ui.ctx().data_mut(|d| d.remove::<DragState>(id));
+                }
+                return;
+            }
+
+            let pointer_pos = self.ui.input(|i| i.pointer.interact_pos());
+            let pointer_down = self.ui.input(|i| i.pointer.primary_down());
+            if !drag_state.active && pointer_down && response.hovered() {
+                let start_x = pointer_pos.map(|p| p.x).unwrap_or(gutter_rect.center().x);
+                drag_state = DragState {
+                    left_width: left_rect.width(),
+                    right_width: right_rect.width(),
+                    start_x,
+                    active: true,
+                };
+                self.ui.ctx().data_mut(|d| d.insert_temp(id, drag_state));
+            }
+
+            if drag_state.active && pointer_down {
+                let delta = pointer_pos.map(|p| p.x - drag_state.start_x).unwrap_or(0.0);
+                let min_left = if left_sidebar { 4.0 } else { 32.0 };
+                let min_right = if right_sidebar { 4.0 } else { 32.0 };
+                let new_left = (drag_state.left_width + delta).max(min_left);
+                let new_right = (drag_state.right_width - delta).max(min_right);
+                if let Some(Tile::Container(Container::Linear(linear))) =
+                    self.tree.tiles.get_mut(container_id)
+                {
+                    let share_left = linear.shares[left_id];
+                    let share_right = linear.shares[right_id];
+                    let share_sum = share_left + share_right;
+                    let current_sidebar_share = if sidebar_on_left {
+                        share_left
+                    } else {
+                        share_right
+                    };
+
+                    let total = new_left + new_right;
+                    let new_left_share = share_sum * new_left / total.max(1.0);
+                    let new_right_share = share_sum - new_left_share;
+                    let left_clamped = new_left_share.max(0.01);
+                    let right_clamped = new_right_share.max(0.01);
+                    self.apply_shares(container_id, left_id, right_id, left_clamped, right_clamped);
+
+                    let new_sidebar_share = if sidebar_on_left {
+                        new_left_share
+                    } else {
+                        new_right_share
+                    };
+                    let mask_threshold = min_sidebar_share * 1.05;
+                    if new_sidebar_share <= mask_threshold && !self.mask_state.masked(sidebar_kind)
+                    {
+                        let prev_last = self.mask_state.last_share(sidebar_kind);
+                        let should_update_last = current_sidebar_share > min_sidebar_share * 1.5;
+                        let store_share = if should_update_last {
+                            Some(current_sidebar_share)
+                        } else {
+                            prev_last
+                        };
+                        self.mask_state.set_last_share(sidebar_kind, store_share);
+                        self.mask_state.set_masked(sidebar_kind, true);
+                    }
+                }
+            }
+
+            if drag_state.active && !pointer_down {
+                self.ui.ctx().data_mut(|d| d.remove::<DragState>(id));
+            }
+        }
+
+        fn process_container(&mut self, container_id: TileId) {
+            let parent_rect = match self.tree.tiles.rect(container_id) {
+                Some(rect) => rect,
+                None => return,
+            };
+
+            let visible_children: Vec<TileId> = match self.tree.tiles.get(container_id) {
+                Some(Tile::Container(Container::Linear(linear))) => linear
+                    .children
+                    .iter()
+                    .copied()
+                    .filter(|child| self.tree.tiles.is_visible(*child))
+                    .collect(),
+                _ => Vec::new(),
+            };
+
+            if visible_children.len() < 2 {
+                return;
+            }
+
+            let mut child_data = Vec::new();
+            for child in &visible_children {
+                if let Some(rect) = self.tree.tiles.rect(*child) {
+                    let sidebar_kind = self.sidebar_kind(*child);
+                    child_data.push((*child, rect, sidebar_kind));
+                }
+            }
+
+            if child_data.len() < 2 {
+                return;
+            }
+
+            for pair in child_data.windows(2) {
+                let (left_id, left_rect, left_kind) = pair[0];
+                let (right_id, right_rect, right_kind) = pair[1];
+                self.process_pair(
+                    container_id,
+                    parent_rect,
+                    left_id,
+                    left_rect,
+                    left_kind,
+                    right_id,
+                    right_rect,
+                    right_kind,
+                );
+            }
+        }
+
+        fn run(mut self) -> Vec<ShareUpdate> {
+            let linear_ids: Vec<_> = self
+                .tree
+                .tiles
+                .iter()
+                .filter_map(|(id, tile)| {
+                    if matches!(tile, Tile::Container(Container::Linear(_))) {
+                        Some(*id)
+                    } else {
+                        None
+                    }
+                })
+                .collect();
+
+            for container_id in linear_ids {
+                self.process_container(container_id);
+            }
+            self.share_updates
+        }
+    }
+
+    let ctx = GutterCtx {
+        tree,
+        ui,
+        painter,
+        gutter_width,
+        mask_state,
+        share_updates: Vec::new(),
+    };
+    ctx.run()
+}
